requirements.txt
appdirs==1.4.3
argcomplete==1.8.2
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
Babel==2.4.0
base58==0.2.4
boto3==1.4.4
botocore==1.5.74
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
clickclick==1.2.1
connexion==1.0.129
constantly==15.1.0
coverage==4.4.1
cryptography==1.9
Delorean==0.6.0
docutils==0.13.1
durationpy==0.4
dynamo3==0.4.10
enum34==1.1.6
Flask==0.12.1
Flask-Cors==3.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
hjson==2.0.7
humanize==0.5.1
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
inflection==0.3.1
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jwcrypto==0.4.0
kappa==0.6.0
lambda-packages==0.15.1
MarkupSafe==1.0
mirakuru==0.8.2
packaging==16.8
pathlib==1.0.1
placebo==0.8.1
pluggy==0.4.0
port-for==0.3.1
psutil==5.2.2
py==1.4.33
pycparser==2.17
PyJWT==1.5.0
pynamodb==3.0.1
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.5.1
pytest-sugar==0.8.0
python-dateutil==2.6.0
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
requests==2.18.1
s3transfer==0.1.10
six==1.10.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
termcolor==1.1.0
toml==0.9.2
tqdm==4.14.0
troposphere==1.9.4
tzlocal==1.4
Unidecode==0.4.20
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.42.1
zope.interface==4.4.2
